+++ ReleaseNotes.md
@@ -1,2 +1,3 @@
-# Release 2.27.1
-* Expose Analysis access endpoints.
+# Release 2.28.0
+* Allow passing fleet context when retrieving analytic sets. 
+* Added FieldSet access endpoints.